Log Out working properly, Need to work on getting the attendance part
