+++ app/src/main/java/com/axyz/upasthithguru/activity/MainActivity.kt
@@ -31,7 +31,6 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
         val currentUser: User? = app.currentUser
         if (currentUser != null && currentUser.loggedIn) {
-            Toast.makeText(this,"User Already Signed In", Toast.LENGTH_SHORT).show()
             startActivity(Intent(this, HomeActivity::class.java))
             finish()
         } else {
