+++ app/src/main/res/layout/fragment_attendance.xml
@@ -44,10 +44,6 @@
             android:layout_below="@+id/upastithLogoImageView"
             android:layout_alignParentStart="true"
             android:layout_alignParentEnd="true"
-            android:layout_marginStart="3dp"
-            android:layout_marginTop="1dp"
-            android:layout_marginEnd="-3dp"
-
             android:background="@drawable/curved_textview"
             android:backgroundTint="@color/white" />
