+++ app/src/main/res/layout/fragment_home.xml
@@ -43,13 +43,8 @@
             android:layout_below="@+id/upastithLogoImageView"
             android:layout_alignParentStart="true"
             android:layout_alignParentEnd="true"
-            android:layout_marginStart="3dp"
-            android:layout_marginTop="1dp"
-            android:layout_marginEnd="3dp"
-
             android:background="@drawable/curved_textview"
             android:backgroundTint="@color/white" />
-
         <TextView
             android:id="@+id/text_your_courses"
             android:layout_width="wrap_content"
