+++ app/src/main/res/layout/fragment_setting.xml
@@ -42,9 +42,6 @@
             android:layout_below="@+id/upastithLogoImageView"
             android:layout_alignParentStart="true"
             android:layout_alignParentEnd="true"
-            android:layout_marginStart="3dp"
-            android:layout_marginTop="1dp"
-            android:layout_marginEnd="-3dp"
             android:background="@drawable/curved_textview"
             android:backgroundTint="@color/white" />
 
@@ -208,6 +205,12 @@
             </RelativeLayout>
 
         </androidx.cardview.widget.CardView>
+        <ProgressBar
+            android:id="@+id/progress_bar"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_centerInParent="true"
+            android:visibility="gone" />
     </RelativeLayout>
